move srouce files, update appropriately
